Saltc: Pull out token and special_char files from tokenizer folders
