+++ saltc/include/special_char.h
