+++ saltc/include/token.h
@@ -4,7 +4,7 @@
 #ifndef TOKEN_H_
 #define TOKEN_H_
 
-#include "../utils.h"
+#include "utils.h"
 #include <string>
 #include <map>
